api/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.8
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
pdfminer.six==20201018
pycodestyle==2.7.0
pycparser==2.20
pytz==2021.1
sortedcontainers==2.4.0
sqlparse==0.4.1
toml==0.10.2
